fix: test 오류 수정

+++ src/main/java/roomescape/service/ReservationService.java
@@ -5,6 +5,7 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.dto.ReservationRequest;
 import roomescape.domain.dto.ReservationResponse;
 import roomescape.domain.entity.Reservation;
+import roomescape.exception.InvalidReservationRequestException;
 import roomescape.repository.ReservationRepository;
 import roomescape.valid.ReservationValidator;
 
@@ -34,6 +35,9 @@ public class ReservationService {
 
     @Transactional
     public void deleteById(Long id) {
+        if (id == null || !reservationRepository.existsById(id)) {
+            throw new InvalidReservationRequestException("이미 삭제된 예약입니다.");
+        }
         reservationRepository.deleteById(id);
     }
 }
